pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "player-profiles"
version = "0.1.0"
description = "Hierarchical Bayesian MLB player projection and pitch-type matchup system"
license = {text = "MIT"}
authors = [{name = "Kekoa Santana", email = "kekoasantana42@gmail.com"}]
requires-python = ">=3.11"
dependencies = [
    "pymc>=5.10",
    "arviz>=0.17",
    "pandas>=2.1",
    "numpy>=1.25",
    "scipy>=1.11",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "scikit-learn>=1.3",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "pyyaml>=6.0",
    "pyarrow>=14.0",
    "joblib>=1.3",
    "xgboost>=2.0",
]

[project.optional-dependencies]
ml = [
    "lightgbm>=4.0",
]
dashboard = [
    "streamlit>=1.30",
]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ipykernel>=6.25",
    "jupyter>=1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "integration: slower end-to-end tests that sample real PyMC models on synthetic data",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
# Start permissive — ratchet gradually
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter eventually)
    "E741",   # ambiguous variable name (common in math-heavy code)
    "F401",   # unused imports (re-exports in __init__.py)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]